+++ changelog.md
@@ -1,1 +1,3 @@
-- Updated to 1.19.1
+- Added directional goat horn sounds
+- Made music disc hearing range configurable
+- Made music disc range in line with vanilla music discs