+++ changelog.md
@@ -2,3 +2,4 @@
 - Made music disc hearing range configurable
 - Made music disc range in line with vanilla music discs
 - Added config options to disallow uploading mp3s or wavs
+- Added volume categories for music disc and goat horn sounds